test and implement api /products/id

+++ app/src/main/java/com/c241ps447/prodswing/data/response/ProductsResponse.kt
@@ -14,7 +14,7 @@ data class ProductsResponseItem(
 	val productID: String? = null,
 
 	@field:SerializedName("price")
-	val price: Int? = null,
+	val price: Double? = null,
 
 	@field:SerializedName("review")
 	val review: String? = null,
